requirements.txt
requests
pathlib
python-crontab